Catch errors when removing listeners within quit()
There's a race condition when calling the engine's `quit` command. Because we
don't have catch block, we were getting this:

    UnhandledPromiseRejectionWarning: Unhandled promise rejection. This error
    originated either by throwing inside of an async function without a catch
    block, or by rejecting a promise which was not handled with .catch().
    (rejection id: 2)

Adding this catch block solves the problem. I suspect `proc` itself is removed
inside the promise above. When `this.proc.removeAllListeners()` is called shortly
after, that's when it crashes.

I couldn't pinpoint to whether it's the proc itself that was deleted or
something else, but that's not too relevant for this commit, as it fixes it in a
way that enable me to use it somewhere else without problems.

This should close https://github.com/ebemunk/node-uci/issues/12